+++ src/alert/AlertActions.tsx
@@ -1,7 +1,6 @@
 import { Box, BoxProps } from "../box";
 import { useTheme } from "../theme";
-import { tcm } from "../utils";
-import { forwardRefWithAs } from "../utils/types";
+import { forwardRefWithAs, tcm } from "../utils";
 
 import { useAlertContext } from "./Alert";