+++ src/box/stories/BoxBasic.stories.tsx
@@ -1,9 +1,9 @@
 import { Meta } from "@storybook/react";
 
-import { Badge, ClockIcon } from "../..";
+import { Badge } from "../../badge";
 import { Button } from "../../button";
-
-import { Box } from "./BoxBasic.component";
+import { ClockIcon } from "../../icons";
+import { Box } from "../index";
 
 export default {
   title: "Primitives/Box/Basic",