+++ src/switch/SwitchInput.tsx
@@ -5,8 +5,7 @@ import {
 
 import { CommonFieldProps, useFormControl } from "../form-field";
 import { useTheme } from "../theme";
-import { tcm } from "../utils";
-import { forwardRefWithAs } from "../utils/types";
+import { forwardRefWithAs, tcm } from "../utils";
 
 import { useSwitchContext } from "./Switch";